Make dropdown components more generic
This will make it easier to build dropdown like components like menu's, etc.
